requirements.txt
praw>=6.0.0
matrix_client>=0.3.2
markdown>=3.0.1
discord.py>=0.16.12
